add filter tipos residuos and test

+++ src/main/java/org/circle8/controller/RecorridoController.java
@@ -23,8 +23,8 @@ public class RecorridoController {
 
 	private final RecorridoResponse mock = RecorridoResponse.builder()
 		.fechaRetiro(ZonedDateTime.of (2023, 1, 1, 16, 30, 0, 0, Dates.UTC))
-		.recicladorId(1).recicladorUri("/reciclador/1")
-		.zonaId(1).zonaUri("/organizacion/1/zona/1")
+		.recicladorId(1L).recicladorUri("/reciclador/1")
+		.zonaId(1L).zonaUri("/organizacion/1/zona/1")
 		.puntoInicio(new PuntoResponse(-34.6347176f,-58.5587959f))
 		.puntoFin(new PuntoResponse(-34.6516556f,-58.5356009f))
 		.puntos(List.of(
@@ -46,7 +46,7 @@ public class RecorridoController {
 	 */
 	public ApiResponse post(Context ctx) {
 		return mock.toBuilder()
-			.zonaId(Integer.parseInt(ctx.pathParam(ID_ZONA_PARAM)))
+			.zonaId(Long.parseLong(ctx.pathParam(ID_ZONA_PARAM)))
 			.zonaUri("/organizacion/"+ctx.pathParam(ID_ORGANIZACION_PARAM)+"/zona/"+ctx.pathParam(ID_ZONA_PARAM))
 			.build();
 	}
@@ -57,7 +57,7 @@ public class RecorridoController {
 	public ApiResponse put(Context ctx) {
 		return mock.toBuilder()
 			.id(Integer.parseInt(ctx.pathParam("id")))
-			.zonaId(Integer.parseInt(ctx.pathParam(ID_ZONA_PARAM)))
+			.zonaId(Long.parseLong(ctx.pathParam(ID_ZONA_PARAM)))
 			.zonaUri("/organizacion/"+ctx.pathParam(ID_ORGANIZACION_PARAM)+"/zona/"+ctx.pathParam(ID_ZONA_PARAM))
 			.build();
 	}
